# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TON Site Builder API",
    "description": "Backend Telegram Mini App для создания и публикации сайтов в сети TON",
    "version": "1.0.0"
  },
  "paths": {
    "/api/user/auth": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Auth",
        "description": "Вход в Mini App: валидация подписи initData, создание пользователя при первом входе.",
        "operationId": "auth_api_user_auth_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/ton-proof-payload": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Ton Proof Payload",
        "description": "Одноразовый nonce для ton_proof.\n\nДополнение к контракту: без серверного nonce подпись кошелька можно\nпереиспользовать, поэтому payload выдаём только мы и только на один раз.",
        "operationId": "ton_proof_payload_api_user_ton_proof_payload_get",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TonProofPayloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/connect-wallet": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Connect Wallet",
        "description": "Привязка кошелька: подтверждаем владение адресом проверкой ton_proof.",
        "operationId": "connect_wallet_api_user_connect_wallet_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectWalletRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectWalletResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/settings": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_user_settings_patch",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Me",
        "operationId": "me_api_user_me_get",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tariffs": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "List Tariffs",
        "operationId": "list_tariffs_api_tariffs_get",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TariffOut"
                  },
                  "title": "Response List Tariffs Api Tariffs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/purchase": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Purchase",
        "description": "Создаёт pending-платёж и отдаёт транзакцию для подписи в TON Connect.",
        "operationId": "purchase_api_subscriptions_purchase_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/confirm": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Confirm",
        "description": "Подтверждение оплаты: проверка транзакции on-chain, затем активация подписки.",
        "operationId": "confirm_api_subscriptions_confirm_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "My Subscriptions",
        "operationId": "my_subscriptions_api_subscriptions_get",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionOut"
                  },
                  "title": "Response My Subscriptions Api Subscriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthRequest": {
        "properties": {
          "init_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init Data"
          }
        },
        "type": "object",
        "title": "AuthRequest"
      },
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          }
        },
        "type": "object",
        "required": [
          "user",
          "is_admin"
        ],
        "title": "AuthResponse"
      },
      "ConfirmPaymentRequest": {
        "properties": {
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          },
          "tx_hash": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "tx_hash"
        ],
        "title": "ConfirmPaymentRequest"
      },
      "ConnectWalletRequest": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "maxLength": 128,
            "minLength": 10,
            "title": "Wallet Address"
          },
          "ton_proof": {
            "$ref": "#/components/schemas/TonProof"
          },
          "public_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Key"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "ton_proof"
        ],
        "title": "ConnectWalletRequest"
      },
      "ConnectWalletResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          }
        },
        "type": "object",
        "required": [
          "success",
          "wallet_address"
        ],
        "title": "ConnectWalletResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Language": {
        "type": "string",
        "enum": [
          "ru",
          "en"
        ],
        "title": "Language"
      },
...
      "PurchaseRequest": {
        "properties": {
          "tariff_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tariff Id"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          }
        },
        "type": "object",
        "required": [
          "tariff_id"
        ],
        "title": "PurchaseRequest"
      },
      "PurchaseResponse": {
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TonConnectTransaction"
          },
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          }
        },
        "type": "object",
        "required": [
          "transaction",
          "payment_id"
        ],
        "title": "PurchaseResponse"
      },
...
      "SubscriptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tariff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TariffOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_forever": {
            "type": "boolean",
            "title": "Is Forever"
          },
          "is_trial": {
            "type": "boolean",
            "title": "Is Trial"
          },
          "status": {
            "$ref": "#/components/schemas/SubscriptionStatus"
          },
          "granted_by_admin": {
            "type": "boolean",
            "title": "Granted By Admin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "starts_at",
          "is_forever",
          "is_trial",
          "status",
          "granted_by_admin"
        ],
        "title": "SubscriptionOut"
      },
      "SubscriptionResponse": {
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionOut"
          }
        },
        "type": "object",
        "required": [
          "subscription"
        ],
        "title": "SubscriptionResponse"
      },
      "SubscriptionStatus": {
        "type": "string",
        "enum": [
          "active",
          "expiring_soon",
          "expired",
          "manual"
        ],
        "title": "SubscriptionStatus"
      },
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
...
      "TariffDuration": {
        "type": "string",
        "enum": [
          "month",
          "3month",
          "6month",
          "12month",
          "forever"
        ],
        "title": "TariffDuration"
      },
      "TariffKind": {
        "type": "string",
        "enum": [
          "base",
          "pro",
          "custom_code"
        ],
        "title": "TariffKind"
      },
      "TariffOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sites_limit": {
            "type": "integer",
            "title": "Sites Limit"
          },
          "duration": {
            "$ref": "#/components/schemas/TariffDuration"
          },
          "price_ton": {
            "type": "string",
            "title": "Price Ton"
          },
          "kind": {
            "$ref": "#/components/schemas/TariffKind"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sites_limit",
          "duration",
          "price_ton",
          "kind",
          "is_active"
        ],
        "title": "TariffOut"
      },
...
      "Theme": {
        "type": "string",
        "enum": [
          "light",
          "dark"
        ],
        "title": "Theme"
      },
...
      "TonConnectTransaction": {
        "properties": {
          "validUntil": {
            "type": "integer",
            "title": "Validuntil"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/TonConnectMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "network": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          }
        },
        "type": "object",
        "required": [
          "validUntil",
          "messages"
        ],
        "title": "TonConnectTransaction"
      },
      "TonProof": {
        "properties": {
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          },
          "domain": {
            "$ref": "#/components/schemas/TonProofDomain"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "state_init": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Init"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "domain",
          "signature",
          "payload"
        ],
        "title": "TonProof"
      },
      "TonProofDomain": {
        "properties": {
          "lengthBytes": {
            "type": "integer",
            "title": "Lengthbytes"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "lengthBytes",
          "value"
        ],
        "title": "TonProofDomain"
      },
      "TonProofPayloadResponse": {
        "properties": {
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "payload",
          "expires_at"
        ],
        "title": "TonProofPayloadResponse"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "telegram_id": {
            "type": "integer",
            "title": "Telegram Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "wallet_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Address"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "theme": {
            "$ref": "#/components/schemas/Theme"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "telegram_id",
          "language",
          "theme",
          "is_admin",
          "created_at"
        ],
        "title": "UserOut"
      },
      "UserSettingsRequest": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Theme"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UserSettingsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}